Handle double-digit version numbers
The underlying SQL queries don't behave as expected when part of the
version number has more than one digit and alphasorting produces wrong
results. Define a comparison function for class Version and use that
to determine what the latest version is.

While touching the code, try to be smarter about which tags we want to
look at by filtering the format beforehand.